fix memberout

+++ src/main/java/controllers/MembersController.java
@@ -163,15 +163,33 @@ public class MembersController extends HttpServlet {
 
 			}else if(cmd.equals("/memberout.members")) { 
 
-
 				String userId = request.getParameter("userId");
-				String userPw = SecurityUtils.sha512(request.getParameter("loginPw"));
+				String userPw = SecurityUtils.sha512(request.getParameter("userPw"));
 				
-				int result = dao.delete(userId, userPw);
+				// ID 체크
+				boolean idCheck = dao.isIdExist(userId);
 				
-				String resp = g.toJson(result);
-				response.getWriter().append(resp);
-
+				if(!idCheck) {
+					response.getWriter().append("1");
+					return;
+				}
+				
+				// 비밀번호 체크
+				boolean pwCheck = dao.isPwExist(userId, userPw);
+				
+				if(!pwCheck) {
+					response.getWriter().append("2");
+					return;
+					
+				}else {
+					int result = dao.delete(userId,userPw);
+					if(result>0) {
+						request.getSession().removeAttribute("userId");
+						request.getSession().removeAttribute("userno");
+						response.getWriter().append("3");
+						return;
+					}
+				}
 
 			}else if(cmd.equals("/mypage.members")) { 
 				String userId = (String)request.getSession().getAttribute("loginID");
